Kernel: Move a bunch of generic devices code into new subdirectory
