+++ Kernel/Arch/init.cpp
@@ -14,18 +14,18 @@
 #include <Kernel/Bus/VirtIO/Device.h>
 #include <Kernel/CommandLine.h>
 #include <Kernel/Devices/Audio/Management.h>
-#include <Kernel/Devices/DeviceControlDevice.h>
 #include <Kernel/Devices/DeviceManagement.h>
-#include <Kernel/Devices/FullDevice.h>
+#include <Kernel/Devices/Generic/DeviceControlDevice.h>
+#include <Kernel/Devices/Generic/FullDevice.h>
+#include <Kernel/Devices/Generic/MemoryDevice.h>
+#include <Kernel/Devices/Generic/NullDevice.h>
+#include <Kernel/Devices/Generic/RandomDevice.h>
+#include <Kernel/Devices/Generic/SelfTTYDevice.h>
+#include <Kernel/Devices/Generic/ZeroDevice.h>
 #include <Kernel/Devices/HID/Management.h>
 #include <Kernel/Devices/KCOVDevice.h>
-#include <Kernel/Devices/MemoryDevice.h>
-#include <Kernel/Devices/NullDevice.h>
 #include <Kernel/Devices/PCISerialDevice.h>
-#include <Kernel/Devices/RandomDevice.h>
-#include <Kernel/Devices/SelfTTYDevice.h>
 #include <Kernel/Devices/SerialDevice.h>
-#include <Kernel/Devices/ZeroDevice.h>
 #include <Kernel/FileSystem/SysFS/Registry.h>
 #include <Kernel/FileSystem/SysFS/Subsystems/Firmware/Directory.h>
 #include <Kernel/FileSystem/VirtualFileSystem.h>
