+++ Kernel/Devices/Audio/Channel.cpp
@@ -7,7 +7,7 @@
 #include <Kernel/API/Ioctl.h>
 #include <Kernel/Devices/Audio/Management.h>
 #include <Kernel/Devices/DeviceManagement.h>
-#include <Kernel/Devices/RandomDevice.h>
+#include <Kernel/Devices/Generic/RandomDevice.h>
 #include <Kernel/Random.h>
 #include <Kernel/Sections.h>
