+++ Kernel/Devices/Generic/ConsoleDevice.h
