+++ Kernel/Devices/Generic/FullDevice.cpp
@@ -6,7 +6,7 @@
 
 #include <Kernel/API/POSIX/errno.h>
 #include <Kernel/Devices/DeviceManagement.h>
-#include <Kernel/Devices/FullDevice.h>
+#include <Kernel/Devices/Generic/FullDevice.h>
 #include <Kernel/Sections.h>
 
 namespace Kernel {
