+++ Kernel/Devices/Generic/NullDevice.h
