+++ Kernel/Devices/Generic/RandomDevice.cpp
@@ -5,7 +5,7 @@
  */
 
 #include <Kernel/Devices/DeviceManagement.h>
-#include <Kernel/Devices/RandomDevice.h>
+#include <Kernel/Devices/Generic/RandomDevice.h>
 #include <Kernel/Random.h>
 #include <Kernel/Sections.h>
