+++ Kernel/Devices/Generic/RandomDevice.h
