+++ Kernel/Devices/Generic/ZeroDevice.h
