+++ Kernel/Random.cpp
@@ -13,7 +13,7 @@
 #elif ARCH(AARCH64)
 #    include <Kernel/Arch/aarch64/ASM_wrapper.h>
 #endif
-#include <Kernel/Devices/RandomDevice.h>
+#include <Kernel/Devices/Generic/RandomDevice.h>
 #include <Kernel/Random.h>
 #include <Kernel/Sections.h>
 #include <Kernel/Time/TimeManagement.h>
