+++ Kernel/TTY/VirtualConsole.h
@@ -10,7 +10,7 @@
 #include <AK/Noncopyable.h>
 #include <AK/Vector.h>
 #include <Kernel/API/KeyCode.h>
-#include <Kernel/Devices/ConsoleDevice.h>
+#include <Kernel/Devices/Generic/ConsoleDevice.h>
 #include <Kernel/Devices/HID/Management.h>
 #include <Kernel/Graphics/Console/Console.h>
 #include <Kernel/TTY/TTY.h>
